+++ SwiftUI/MapkitEx/MapkitEx/View/ContentView.swift
@@ -1,0 +1,20 @@
+//
+//  ContentView.swift
+//  MapkitEx
+//
+//  Created by jaewon Lee on 4/28/25.
+//
+
+import SwiftUI
+import MapKit
+
+struct ContentView: View {
+    var body: some View {
+        MapView()
+    }
+}
+
+#Preview {
+    ContentView()
+}
+
